fix(helia): avoid 64-bit shape division
Use the target word size for pre-multiplication element bounds so Cortex-M builds do not pull in 64-bit division helpers.

Refs #283

+++ tensorflow/lite/micro/kernels/helia/helia_data_movement.h
@@ -57,17 +57,17 @@ inline TfLiteStatus HeliaDataMovementElements(TfLiteContext* context,
                                               const TfLiteIntArray* dims,
                                               TfLiteType type, int32_t* count) {
   TF_LITE_ENSURE(context, dims != nullptr && dims->size >= 0);
-  uint64_t elements = 1;
+  size_t elements = 1;
   const size_t width = type == kTfLiteFloat16 ? 2 : 4;
-  const uint64_t max_elements_by_width =
+  const size_t max_elements_by_width =
       std::numeric_limits<size_t>::max() / width;
-  const uint64_t max_elements =
-      max_elements_by_width < static_cast<uint64_t>(INT32_MAX)
+  const size_t max_elements =
+      max_elements_by_width < static_cast<size_t>(INT32_MAX)
           ? max_elements_by_width
-          : static_cast<uint64_t>(INT32_MAX);
+          : static_cast<size_t>(INT32_MAX);
   for (int i = 0; i < dims->size; ++i) {
     TF_LITE_ENSURE(context, dims->data[i] >= 0);
-    const uint64_t extent = dims->data[i];
+    const size_t extent = dims->data[i];
     TF_LITE_ENSURE(context, extent == 0 || elements <= max_elements / extent);
     elements *= extent;
   }
